refactor(precompiles): extract b20_security business logic into SecurityManagement trait
Mirrors the shared B-20 capability-trait pattern (Burnable, Mintable,
Pausable, Transferable, Permittable, Configurable, RoleManaged). All
security-specific business operations move from inherent methods on
B20SecurityToken inside dispatch.rs into a new SecurityManagement trait
under b20_security/management.rs:

- Associated consts: SECURITY_OPERATOR_ROLE, BURN_FROM_ROLE, REDEEM_SENDER_POLICY
- Default-impl methods: ensure_security_operator, ensure_burn_from_role,
  to_shares, security_redeem(_with_memo)?, security_redeem_burn,
  emit_redeemed, batch_mint, batch_burn, announce
- Required methods (state on the concrete token): is_announcement_active,
  begin_announcement, dispatch_internal_call (forwards to inner_with_privilege)

dispatch.rs is now thinned to ABI decode plus thin trait calls. The
match arms in handle_security_call are unchanged — Self::SECURITY_OPERATOR_ROLE
etc. resolve through trait projection now that SecurityManagement is in scope.

No behavior change.
